// src/app/models/admin/baocao/tailieudinhkem.model.ts
export class OutputTaiLieuModel {
    idtailieu: string;
    tentailieu: string;
    filedinhkem: string;
    duongdan: string;
    thutu: string;
    dungluong: number;
    dinhdang: number;
    idbaocao: string;
    tenBaoCao: string;
}

export class InputTaiLieuModel {
    idtailieu: string;
    tentailieu: string;
    filedinhkem: string;
    duongdan: string;
    thutu: string;
    dungluong: number;
    dinhdang: number;
    idbaocao: string;
}